SampleChainsDynamicHMC

SampleChainsDynamicHMC.dynamichmc

SampleChainsDynamicHMC.dynamichmc — Method

dynamichmc(
  ; init          = ()
  , warmup_stages = DynamicHMC.default_warmup_stages()
  , algorithm     = DynamicHMC.NUTS()
  , reporter      = DynamicHMC.NoProgressReport()
)

init: a NamedTuple that can contain the following fields (all of them optional and provided with reasonable defaults):

q: initial position. Default: random (uniform [-2,2] for each coordinate).

κ: kinetic energy specification. Default: Gaussian with identity matrix.

ϵ: a scalar for initial step size, or nothing for heuristic finders.

warmup_stages: a sequence of warmup stages. See DynamicHMC.default_warmup_stages and DynamicHMC.fixed_stepsize_warmup_stages; the latter requires an ϵ in initialization.

algorithm: see DynamicHMC.NUTS. It is very unlikely you need to modify this, except perhaps for the maximum depth.

reporter: how progress is reported. This is currently silent by default (see DynamicHMC.NoProgressReport), but this default will likely change in future releases.

For more details see https://tamaspapp.eu/DynamicHMC.jl/stable/interface/

Example

julia> using LinearAlgebra

julia> config = dynamichmc(
           warmup_stages=default_warmup_stages(
               M=Symmetric, # adapt dense positive definite metric
               stepsize_adaptation=DualAveraging(δ=0.9), # target acceptance rate 0.9
               doubling_stages=7, # 7-stage metric adaptation
           ),
           reporter=LogProgressReport(), # log progress using Logging
       );